% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_op_utils.R
\name{\%inn\%}
\alias{\%inn\%}
\title{\%inn\%}
\usage{
x \%inn\% table
}
\arguments{
\item{x}{character to check if it exists in a vector or list}

\item{table}{list, data frame or any vector}
}
\value{
bool value TRUE if it exists FALSE if it does not
}
\description{
\%inn\%
}
\examples{
.check = "a" \%inn\% data.frame( a = 1 : 5 )
}
