% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag_df.R
\name{lag_df}
\alias{lag_df}
\title{lag_df
Creates extra columns given as a list of column names and sequences.
useful to have multiple selection of lags of some columns while some columns to
be present only one or two lags or none.}
\usage{
lag_df(df, laglist)
}
\arguments{
\item{df}{data.frame or tibble}

\item{laglist}{a list of column names as index and lag sequence as value}
}
\value{
tibble
}
\description{
lag_df
Creates extra columns given as a list of column names and sequences.
useful to have multiple selection of lags of some columns while some columns to
be present only one or two lags or none.
}
\examples{
df <-data.frame( a = 1:15 , b = 2:16 )
tb <- lag_df( df , laglist= list(a = 1:5 , b = 1:3  ) )

}
