% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag_df.R
\name{lag_df2}
\alias{lag_df2}
\title{lag_df2
Creates extra columns given as a list of column names and sequences.
useful to have multiple selection of lags of some columns while some columns to
be present only one or two lags or none.}
\usage{
lag_df2(df, laglist)
}
\arguments{
\item{df}{data.frame or tibble}

\item{laglist}{a list of column names as index and lag sequence as value}
}
\value{
data.frame
}
\description{
lag_df2
Creates extra columns given as a list of column names and sequences.
useful to have multiple selection of lags of some columns while some columns to
be present only one or two lags or none.
}
\examples{
lag_df2( data.frame(list( a = 1: 5 ) , list(a = 1: 2 )) )

}
