% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{print.vyos_GETPREP}
\alias{print.vyos_GETPREP}
\title{print.vyos_GETPREP
Generic method for S3 vyos_GETPREP object}
\usage{
\method{print}{vyos_GETPREP}(x, ...)
}
\arguments{
\item{x}{S3 vyos_GETPREP object}

\item{...}{further arguments passed to or from other methods.}
}
\value{
S3 vyos_GETPREP object
}
\description{
print.vyos_GETPREP
Generic method for S3 vyos_GETPREP object
}
\examples{
\dontrun{

obj <- get_series(template_test())
print(obj)
}
}
