% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_df_utils.R
\name{remove_columns}
\alias{remove_columns}
\title{Remove a column or columns from data.frame}
\usage{
remove_columns(df, column_names , verbose = FALSE  )
}
\arguments{
\item{df}{data.frame or tibble}

\item{column_names}{column name or column names as a character vector}

\item{verbose}{bool gives extra info when removed a column.}
}
\value{
data.frame
}
\description{
Remove a column or columns from data.frame
}
\examples{
 df <- remove_columns(cars, "speed")

}
