% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_df_utils.R
\name{remove_na_safe}
\alias{remove_na_safe}
\title{remove_na_safe}
\usage{
remove_na_safe(df , verbose = FALSE )
}
\arguments{
\item{df}{data.frame to remove na rows from the beginning and from the end}

\item{verbose}{give detailed info while removing NA values}
}
\value{
data.frame
}
\description{
removes rows from the beginning and end of data.frame provided that all columns
are NA.
}
\examples{

df <- data.frame( a = c( NA, 2: 7 , NA ) ,
                  b = c( NA, NA, 5, NA, 12, NA, 8, 9 ) )
df2 <- remove_na_safe( df )
}
