% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_series.R
\name{get_series}
\alias{get_series}
\title{Requests data from multiple data sources.
EDDS API and FRED API at this version.}
\usage{
get_series(
  index = null,
  start_date = default_start_date(),
  end_date = default_end_date(),
  freq = null,
  cache = FALSE,
  na.remove = TRUE,
  verbose = FALSE,
  ...,
  source = c("multi", "evds", "fred"),
  base = c("multi", "series", "table"),
  debug = FALSE
)
}
\arguments{
\item{index}{character  vector or string}

\item{start_date}{limits the start date of data}

\item{end_date}{limits the end date of data}

\item{freq}{frequency rarely needed}

\item{cache}{cache option if false new request will be made}

\item{na.remove}{it will safely remove NA values only if all columns are NA}

\item{verbose}{if TRUE it prints some information during the process. If FALSE
silently does its job. Gives warning only if something goes wrong.}

\item{...}{for future versions}

\item{source}{source such as evds or fred for internal use for at this version}

\item{base}{table or series on the source for internal use for at this version}

\item{debug}{debug option for development}
}
\value{
vyos_GETPREP is an S3 object which has some generic functions such as print
and excel. \code{print} generic gives some hints to user in order to use requested data
such as creating output with excel function or examining requested data in global
environment.
}
\description{
The get_series() function requests data from data sources.
when more than one index was given as a character vector or a string template
it requests all items individually from corresponding sources by figuring out
from the format of item given.
Combines a data frame when there is a common frequencies. Returns all data
both as a combined data frame and also individual data frames.
}
\examples{
\dontrun{
o <- get_series(template_test())
excel(o)
object <- get_series("UNRATE", start_date = "2000/01/01", na.remove = TRUE)
excel(object)
}
}
