% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag_df.R
\name{lag_df2}
\alias{lag_df2}
\title{lag_df2
Creates extra columns given as a list of column names and sequences.
useful to have multiple selection of lags of some columns while some columns to
be present only one or two lags or none.}
\usage{
lag_df2(df, laglist)
}
\arguments{
\item{df}{data.frame or tibble}

\item{laglist}{a list of column names as index and lag sequence as value}
}
\value{
data.frame
}
\description{
lag_df2
Creates extra columns given as a list of column names and sequences.
useful to have multiple selection of lags of some columns while some columns to
be present only one or two lags or none.
}
\examples{
df <- data.frame(a = 1:15, b = 2:16)
df2 <- lag_df2(df, laglist = list(a = 1:5, b = 1:3))
}
